added func to check the database and moved README file contents

+++ dataAccess/microservicesDAO.go
@@ -2,6 +2,7 @@ package dataAccess
 
 import (
 	"database/sql"
+	"fmt"
 	"log"
 	"os"
 
@@ -15,8 +16,21 @@ func ConnectToDB() {
 		log.Fatal(err)
 	}
 	defer database.Close()
-	// We will test this once we set up a sample GitHub! :D
-	repoLink := "https://github.com/JaclynW/testRepo.git" // Example GitHub repository link.
+
+	var microName string
+	fmt.Print("Enter your Microservice name: ")
+	if _, err := fmt.Scanln(&microName); err != nil {
+		fmt.Println("Error reading input:", err)
+		return
+	}
+
+	/*
+		We will test this once we set up a sample GitHub! :D
+		To test this we will use a pre-made repository but repoLink will be provided through the UI
+		The Developer would provide their repoLink and we would clone it and run the SQL file
+	*/
+	repoLink := "https://github.com/JaclynW/testRepo.git"
+
 	// Temporary local path to clone and store the repository's contents.
 	clonePath := "./tempRepo"
 
@@ -27,10 +41,20 @@ func ConnectToDB() {
 	// At this point, you would typically read this file, parse its contents,
 	// and execute any relevant SQL commands against your database.
 
-	// ... Code to process data and insert into the database ...
+	// Before adding the microservice, check if it already exists in the database.
+	if exists := checkMicroserviceExists(database, microName, repoLink); exists {
+		fmt.Println("Microservice with the provided name and repository link already exists in the database.")
+		return
+	}
 
 	// Insert the microservice's information into the database.
-	addMicroservice(database, "MicroserviceName", repoLink) // MicroserviceName changed to service hook or key?
+	/*
+		For our testRepo, the microservice name is "services"
+		We will ask user to input the microservice name which is stored in microName
+
+		We will eventually need to figure out how to grab teh microservice name JUST from the repoLink
+	*/
+	addMicroservice(database, microName, repoLink)
 
 	// Cleanup: Remove the cloned repository from local storage to free up space.
 	os.RemoveAll(clonePath)
@@ -56,4 +80,20 @@ func addMicroservice(database *sql.DB, name, repoLink string) {
 	// Prepare an SQL statement for insertion.
 	statement, _ := database.Prepare("INSERT INTO microservices (build_script, placeholder) VALUES (?, ?)")
 	statement.Exec(name, repoLink) // Execute the statement with provided values.
+}
+
+// checkMicroserviceExists function checks if a microservice with the given name
+// and repository link already exists in the database.
+func checkMicroserviceExists(database *sql.DB, name, repoLink string) bool {
+	// Prepare a SELECT query to find the microservice with the given name and repo link.
+	row := database.QueryRow("SELECT count(*) FROM microservices WHERE build_script = ? AND placeholder = ?", name, repoLink)
+
+	var count int
+	if err := row.Scan(&count); err != nil {
+		log.Fatalf("Error querying the database: %v", err)
+		return false
+	}
+
+	// Return true if count is more than 0, indicating the microservice exists.
+	return count > 0
 }
